tv_add_usec: fixed handling of > 2^31 ms offset
+ improved a bit

+++ src/tv.h
@@ -52,7 +52,7 @@ uint32_t get_local_mediatime(void);
 double   tv_diff(struct timeval curr_time, struct timeval prev_time);
 uint32_t tv_diff_usec(struct timeval curr_time, struct timeval prev_time);
 void     tv_add(struct timeval *ts, double offset_secs);
-void     tv_add_usec(struct timeval *ts, double offset);
+void     tv_add_usec(struct timeval *ts, long long offset);
 int      tv_gt(struct timeval a, struct timeval b);
 uint32_t get_std_audio_local_mediatime(double samples, int rate);
 uint32_t get_std_video_local_mediatime(void);
